Improve jacobi stats processing
Now when J(X, h*2^n-1) == 0 is found, we just
issue a single '0' on the Jacobi string line.
This is because when X < h*2^n-1 (and here X
is something like a 3 digit number), then
we know that gcd((X, h*2^n-1) != 1 and so
h*2^n-1 is not prime, and thus the stats for
finding v(1) for that value is not interesting.

Added other minor cleanup in stats processing.